Fixed .gitignore and added testing file
got the syntax right for the .gitignore so that backup rkt files are
ignored, and the entire "compiled" folder is ignored (idk why this
exists). Also made the testing.rkt file. This file loads all the
individual test case files and then creates an assert function and then
asserts each test case.